--- DriverPortal/app/src/main/java/com/edulog/driverportal/data/entity/DriverEntity.java
@@ -1,4 +1,0 @@
-package com.edulog.driverportal.data.entity;
-
-public class DriverEntity {
-}
