--- DriverPortal/app/src/main/java/com/edulog/driverportal/domain/executor/PostExecutionThread.java
@@ -1,7 +1,0 @@
-package com.edulog.driverportal.domain.executor;
-
-import io.reactivex.Scheduler;
-
-public interface PostExecutionThread {
-    Scheduler getScheduler();
-}
