--- DriverPortal/app/src/main/java/com/edulog/driverportal/domain/service/AuthService.java
@@ -1,9 +1,0 @@
-package com.edulog.driverportal.domain.service;
-
-import io.reactivex.Observable;
-import okhttp3.ResponseBody;
-import retrofit2.Response;
-
-public interface AuthService {
-    Observable<Response<ResponseBody>> changePassword(String driverId, String oldPassword, String newPassword);
-}
