--- DriverPortal/app/src/main/java/com/edulog/driverportal/presentation/BaseActivity.java
@@ -1,54 +1,0 @@
-package com.edulog.driverportal.presentation;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-
-public abstract class BaseActivity extends AppCompatActivity implements BaseView {
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-
-        if (getPresenter() != null && getView() != null) {
-            // noinspection unchecked
-            getPresenter().attach(getView());
-        }
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-
-        if (getPresenter() != null) {
-            getPresenter().detach();
-        }
-    }
-
-    @Override
-    public void showProgress() {
-
-    }
-
-    @Override
-    public void hideProgress() {
-
-    }
-
-    @Override
-    public void showError(String message) {
-
-    }
-
-    protected BasePresenter getPresenter() {
-        return null;
-    }
-
-    protected BaseView getView() {
-        return null;
-    }
-}
