--- DriverPortal/app/src/main/java/com/edulog/driverportal/presentation/BaseFragment.java
@@ -1,47 +1,0 @@
-package com.edulog.driverportal.presentation;
-
-import android.support.v4.app.Fragment;
-
-public abstract class BaseFragment extends Fragment implements BaseView {
-    @Override
-    public void onStart() {
-        super.onStart();
-
-        if (getPresenter() != null && getViewLayer() != null) {
-            // noinspection unchecked
-            getPresenter().attach(getViewLayer());
-        }
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-
-        if (getPresenter() != null) {
-            getPresenter().detach();
-        }
-    }
-
-    @Override
-    public void showProgress() {
-
-    }
-
-    @Override
-    public void hideProgress() {
-
-    }
-
-    @Override
-    public void showError(String message) {
-
-    }
-
-    protected BasePresenter getPresenter() {
-        return null;
-    }
-
-    protected BaseView getViewLayer() {
-        return null;
-    }
-}
