--- DriverPortal/app/src/main/java/com/edulog/driverportal/presentation/BasePresenter.java
@@ -1,9 +1,0 @@
-package com.edulog.driverportal.presentation;
-
-public interface BasePresenter<V extends BaseView> {
-    void attach(V view);
-
-    void detach();
-
-    void onError(String message);
-}
