--- DriverPortal/app/src/main/java/com/edulog/driverportal/presentation/BaseView.java
@@ -1,9 +1,0 @@
-package com.edulog.driverportal.presentation;
-
-public interface BaseView {
-    void showProgress();
-
-    void hideProgress();
-
-    void showError(String message);
-}
