--- DriverPortal/app/src/main/java/com/edulog/driverportal/presentation/UIThread.java
@@ -1,13 +1,0 @@
-package com.edulog.driverportal.presentation;
-
-import com.edulog.driverportal.domain.executor.PostExecutionThread;
-
-import io.reactivex.Scheduler;
-import io.reactivex.android.schedulers.AndroidSchedulers;
-
-public class UIThread implements PostExecutionThread {
-    @Override
-    public Scheduler getScheduler() {
-        return AndroidSchedulers.mainThread();
-    }
-}
