--- DriverPortal/app/src/main/java/com/edulog/driverportal/presentation/settings/SettingsFragment.java
@@ -1,44 +1,0 @@
-package com.edulog.driverportal.presentation.settings;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.edulog.driverportal.R;
-import com.edulog.driverportal.presentation.BaseFragment;
-
-public class SettingsFragment extends BaseFragment {
-    public static SettingsFragment newInstance() {
-        return new SettingsFragment();
-    }
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Nullable
-    @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View root = inflater.inflate(R.layout.fragment_settings, container, false);
-        return root;
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-
-        setTitle(getResources().getString(R.string.app_name));
-    }
-
-    private void setTitle(String title) {
-        ActionBar actionBar = ((AppCompatActivity) getActivity()).getSupportActionBar();
-        if (actionBar != null) {
-            actionBar.setTitle(title);
-        }
-    }
-}
