--- DriverPortal/app/src/main/java/com/edulog/driverportal/presentation/settings/changepassword/ChangePasswordPresenter.java
@@ -1,14 +1,0 @@
-package com.edulog.driverportal.presentation.settings.changepassword;
-
-import com.edulog.driverportal.presentation.BasePresenter;
-
-import io.reactivex.Observable;
-
-public interface ChangePasswordPresenter extends BasePresenter<ChangePasswordView> {
-    void changePassword(String driverId, String oldPassword, String newPassword);
-
-    void validateUserInput(Observable<CharSequence> driverIdObservable,
-                           Observable<CharSequence> oldPasswordObservable,
-                           Observable<CharSequence> newPasswordObservable,
-                           Observable<CharSequence> confirmNewPasswordObservable);
-}
