+++ DriverPortal/app/src/main/java/com/edulog/driverportal/settings/BasePresenter.java
@@ -1,0 +1,9 @@
+package com.edulog.driverportal.settings;
+
+public interface BasePresenter<V extends BaseView> {
+    void attach(V view);
+
+    void detach();
+
+    void onError(String message);
+}
