+++ DriverPortal/app/src/main/java/com/edulog/driverportal/settings/BaseView.java
@@ -1,0 +1,9 @@
+package com.edulog.driverportal.settings;
+
+public interface BaseView {
+    void showProgress();
+
+    void hideProgress();
+
+    void showError(String message);
+}
