+++ DriverPortal/app/src/main/java/com/edulog/driverportal/settings/UIThread.java
@@ -1,0 +1,13 @@
+package com.edulog.driverportal.settings;
+
+import com.edulog.driverportal.settings.changepassword.domain.executor.PostExecutionThread;
+
+import io.reactivex.Scheduler;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+
+public class UIThread implements PostExecutionThread {
+    @Override
+    public Scheduler getScheduler() {
+        return AndroidSchedulers.mainThread();
+    }
+}
