+++ DriverPortal/app/src/main/java/com/edulog/driverportal/settings/changepassword/data/entity/DriverEntity.java
@@ -1,0 +1,4 @@
+package com.edulog.driverportal.settings.changepassword.data.entity;
+
+public class DriverEntity {
+}
