+++ DriverPortal/app/src/main/java/com/edulog/driverportal/settings/changepassword/domain/executor/PostExecutionThread.java
@@ -1,0 +1,7 @@
+package com.edulog.driverportal.settings.changepassword.domain.executor;
+
+import io.reactivex.Scheduler;
+
+public interface PostExecutionThread {
+    Scheduler getScheduler();
+}
