+++ DriverPortal/app/src/main/java/com/edulog/driverportal/settings/changepassword/domain/service/AuthService.java
@@ -1,0 +1,9 @@
+package com.edulog.driverportal.settings.changepassword.domain.service;
+
+import io.reactivex.Observable;
+import okhttp3.ResponseBody;
+import retrofit2.Response;
+
+public interface AuthService {
+    Observable<Response<ResponseBody>> changePassword(String driverId, String oldPassword, String newPassword);
+}
