--- DriverPortal/app/src/test/java/com/edulog/driverportal/domain/interactor/ChangePasswordUseCaseTest.java
@@ -1,16 +1,0 @@
-package com.edulog.driverportal.domain.interactor;
-
-import org.junit.Test;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-
-public class ChangePasswordUseCaseTest {
-    @Test
-    public void execute_changePasswordShouldBeCalled() {
-        String driverId = "1234";
-        String oldPassword = "old_pass";
-        String newPassword = "new_pass";
-
-    }
-}
