+++ DriverPortal/app/src/test/java/com/edulog/driverportal/settings/changepassword/ChangePasswordPresenterTest.java
@@ -1,0 +1,16 @@
+package com.edulog.driverportal.settings.changepassword;
+
+import org.junit.Test;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+
+public class ChangePasswordPresenterTest {
+    @Test
+    public void changePassword_executeShouldBeCalled() {
+        String driverId = "1234";
+        String oldPassword = "old_pass";
+        String newPassword = "new_pass";
+        String confirmNewPassword = "new_pass";
+    }
+}
